refactor: assorted cleanup
